only copy if exists in upload.sh

+++ upload.sh
@@ -2,8 +2,10 @@
 
 set -ev
 
-# copy latest version of the learning progression from box
-cp ~/box/Learning\ Progressions\ for\ Partial\ Derivatives/Learning\ Progression\ Database.xlsx progression.xlsx
+if test -e ~/box/Learning\ Progressions\ for\ Partial\ Derivatives/Learning\ Progression\ Database.xlsx; then
+    # copy latest version of the learning progression from box
+    cp ~/box/Learning\ Progressions\ for\ Partial\ Derivatives/Learning\ Progression\ Database.xlsx progression.xlsx
+fi
 
 fac
